Fix: quicklinks button style not transferred
Fixes a bug that the styles of quick link button and its dropdown would not be transmitted to the detached window.
I tried at first to inject the styles, but it was a lot, so decided to show (render) them even when no URLs are written in slides. 
"disabled = true" did not allow rendering the dropdown, so I need to put a dummy item to the dropdown list.